big bug with lights and shadow plus nan ray if dir == 1 0 0

+++ includes/libmath.h
@@ -6,7 +6,7 @@
 /*   By: tamigore <tamigore@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/14 18:31:48 by tamigore          #+#    #+#             */
-/*   Updated: 2022/04/14 17:16:16 by tamigore         ###   ########.fr       */
+/*   Updated: 2022/05/02 13:23:25 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,9 +34,9 @@
 
 typedef struct	s_v3
 {
-	float		x;
-	float		y;
-	float		z;
+	double		x;
+	double		y;
+	double		z;
 }				t_v3;
 
 /*
@@ -47,10 +47,10 @@ typedef struct	s_v3
 
 typedef struct	s_vec
 {
-	float		x;
-	float		y;
-	float		z;
-	float		w;
+	double		x;
+	double		y;
+	double		z;
+	double		w;
 }				t_vec;
 
 /*
@@ -65,18 +65,18 @@ typedef struct	s_vec
 
 typedef struct	t_mat
 {
-	float		mat[4][4];
+	double		mat[4][4];
 }				t_mat;
 
 /*
 ** VECTOR.C vector basics
 */
 
-t_vec			v_init(float x, float y, float z, float w);
+t_vec			v_init(double x, double y, double z, double w);
 t_vec			v_add(t_vec a, t_vec b);
 t_vec			v_sub(t_vec a, t_vec b);
 t_vec			v_cross(t_vec a, t_vec b);
-t_vec			v_scale(float x, t_vec a);
+t_vec			v_scale(double x, t_vec a);
 
 /*
 ** VECTOR_2.C vector operation
@@ -84,25 +84,25 @@ t_vec			v_scale(float x, t_vec a);
 
 t_vec			v_prod(t_vec a, t_vec b);
 int				v_comp(t_vec a, t_vec b);
-float			v_len(t_vec v);
+double			v_len(t_vec v);
 t_vec			v_norm(t_vec vec);
-float			v_dot(t_vec a, t_vec b);
+double			v_dot(t_vec a, t_vec b);
 
 /*
 ** ANGLES.C trigo basics
 */
 
-float			deg2rad(float deg);
-float			rad2deg(float rad);
-float			v_cos(t_vec a, t_vec b);
-void			swap(float *x, float *y);
+double			deg2rad(double deg);
+double			rad2deg(double rad);
+double			v_cos(t_vec a, t_vec b);
+void			swap(double *x, double *y);
 
 /*
 ** MATRIX.C matrix init and fill
 */
 
 t_mat	identity_mat_init(void);
-t_mat	scale_mat_init(float scale);
+t_mat	scale_mat_init(double scale);
 t_mat	translation_mat_init(t_vec tranvec);
 
 /*
@@ -116,11 +116,11 @@ t_vec	mat_mult_vec(t_mat a, t_vec b);
 ** MATRIX_TR.C matrix transformation and rotation
 */
 
-t_mat	mat_axis_angle_get(t_vec axis, float cos, float sin, float t);
-t_mat	rotmat_axis_angle(t_vec axis, float angle);
-t_mat	rotx_mat_init(float degrees);
-t_mat	roty_mat_init(float degrees);
-t_mat	rotz_mat_init(float degrees);
+t_mat	mat_axis_angle_get(t_vec axis, double cos, double sin, double t);
+t_mat	rotmat_axis_angle(t_vec axis, double angle);
+t_mat	rotx_mat_init(double degrees);
+t_mat	roty_mat_init(double degrees);
+t_mat	rotz_mat_init(double degrees);
 
 /*
 ** MATH.C
