+++ lib/libmath/srcs/angles.c
@@ -6,30 +6,30 @@
 /*   By: tamigore <tamigore@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 12:15:16 by tamigore          #+#    #+#             */
-/*   Updated: 2022/04/14 17:22:00 by tamigore         ###   ########.fr       */
+/*   Updated: 2022/05/02 13:23:25 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libmath.h"
 
-float	deg2rad(float deg)
+double	deg2rad(double deg)
 {
 	return (deg * M_PI / FLAT_ANGLE);
 }
 
-float	rad2deg(float rad)
+double	rad2deg(double rad)
 {
 	return (rad * FLAT_ANGLE / M_PI);
 }
 
-float	v_cos(t_vec a, t_vec b)
+double	v_cos(t_vec a, t_vec b)
 {
 	return (v_dot(a, b) / v_len(a) * v_len(b));
 }
 
-void		swap(float *x, float *y)
+void		swap(double *x, double *y)
 {
-	float	tmp;
+	double	tmp;
 
 	tmp = *x;
 	*x = *y;
